Mock all external dependencies for doc gen
ReadTheDocs had issues with installing dependencies which
required additional C headers present in the environment.

As there is no need for external dependencies to be present
while generating documentation, all external dependencies are
mocked, so that they stand Sphinx's import.

Additionally, the package is not installed in the environment
to be used for generating documentation, in order to be consisent
with the settings on ReadTheDocs.